+++ examples/react/tanstack-react-query/src/use-graphql.ts
@@ -14,7 +14,7 @@ export function useGraphQL<TResult, TVariables>(
       variables,
     ] as const,
     async ({ queryKey }) => {
-      return fetch('https://swapi-graphql.netlify.app/.netlify/functions/index', {
+      return fetch('https://graphql.org/graphql/', {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',