+++ packages/presets/client/tests/fixtures/enum.ts
@@ -1,0 +1,12 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
+//@ts-ignore
+import gql from 'gql-tag';
+
+//@ts-ignore
+const A = gql(/* GraphQL */ `
+  query FavoriteColor {
+    thing {
+      color
+    }
+  }
+`);