+++ packages/utils/plugins-helpers/src/federation.ts
@@ -35,7 +35,7 @@ export const federationSpec = parse(/* GraphQL */ `
 export function addFederationReferencesToSchema(schema: GraphQLSchema): GraphQLSchema {
   return mapSchema(schema, {
     [MapperKind.OBJECT_TYPE]: type => {
-      if (isFederationObjectType(type, schema)) {
+      if (checkObjectTypeFederationDetails(type, schema)) {
         const typeConfig = type.toConfig();
         typeConfig.fields = {
           [resolveReferenceFieldName]: {
@@ -82,10 +82,15 @@ export function removeFederation(schema: GraphQLSchema): GraphQLSchema {
 
 const resolveReferenceFieldName = '__resolveReference';
 
+interface TypeMeta {
+  hasResolveReference: boolean;
+}
+
 export class ApolloFederation {
   private enabled = false;
   private schema: GraphQLSchema;
   private providesMap: Record<string, string[]>;
+  protected meta: { [typename: string]: TypeMeta } = {};
 
   constructor({ enabled, schema }: { enabled: boolean; schema: GraphQLSchema }) {
     this.enabled = enabled;
@@ -130,7 +135,7 @@ export class ApolloFederation {
    * @param data
    */
   skipField({ fieldNode, parentType }: { fieldNode: FieldDefinitionNode; parentType: GraphQLNamedType }): boolean {
-    if (!this.enabled || !isObjectType(parentType) || !isFederationObjectType(parentType, this.schema)) {
+    if (!this.enabled || !isObjectType(parentType) || !checkObjectTypeFederationDetails(parentType, this.schema)) {
       return false;
     }
 
@@ -158,12 +163,16 @@ export class ApolloFederation {
     if (
       this.enabled &&
       isObjectType(parentType) &&
-      isFederationObjectType(parentType, this.schema) &&
       (isTypeExtension(parentType, this.schema) || fieldNode.name.value === resolveReferenceFieldName)
     ) {
-      const keys = getDirectivesByName('key', parentType);
+      const objectTypeFederationDetails = checkObjectTypeFederationDetails(parentType, this.schema);
+      if (!objectTypeFederationDetails) {
+        return parentTypeSignature;
+      }
 
-      if (keys.length) {
+      const { resolvableKeyDirectives } = objectTypeFederationDetails;
+
+      if (resolvableKeyDirectives.length) {
         const outputs: string[] = [`{ __typename: '${parentType.name}' } &`];
 
         // Look for @requires and see what the service needs and gets
@@ -171,7 +180,7 @@ export class ApolloFederation {
         const requiredFields = this.translateFieldSet(merge({}, ...requires), parentTypeSignature);
 
         // @key() @key() - "primary keys" in Federation
-        const primaryKeys = keys.map(def => {
+        const primaryKeys = resolvableKeyDirectives.map(def => {
           const fields = this.extractFieldSet(def);
           return this.translateFieldSet(fields, parentTypeSignature);
         });
@@ -190,6 +199,13 @@ export class ApolloFederation {
     }
 
     return parentTypeSignature;
+  }
+
+  setMeta(typename: string, update: Partial<TypeMeta>): void {
+    this.meta[typename] = { ...(this.meta[typename] || { hasResolveReference: false }), ...update };
+  }
+  getMeta() {
+    return this.meta;
   }
 
   private isExternalAndNotProvided(fieldNode: FieldDefinitionNode, objectType: GraphQLObjectType): boolean {
@@ -275,7 +291,10 @@ export class ApolloFederation {
  * Checks if Object Type is involved in Federation. Based on `@key` directive
  * @param node Type
  */
-function isFederationObjectType(node: ObjectTypeDefinitionNode | GraphQLObjectType, schema: GraphQLSchema): boolean {
+export function checkObjectTypeFederationDetails(
+  node: ObjectTypeDefinitionNode | GraphQLObjectType,
+  schema: GraphQLSchema
+): { resolvableKeyDirectives: readonly DirectiveNode[] } | false {
   const {
     name: { value: name },
     directives,
@@ -284,9 +303,24 @@ function isFederationObjectType(node: ObjectTypeDefinitionNode | GraphQLObjectTy
   const rootTypeNames = getRootTypeNames(schema);
   const isNotRoot = !rootTypeNames.has(name);
   const isNotIntrospection = !name.startsWith('__');
-  const hasKeyDirective = directives.some(d => d.name.value === 'key');
+  const keyDirectives = directives.filter(d => d.name.value === 'key');
 
-  return isNotRoot && isNotIntrospection && hasKeyDirective;
+  const check = isNotRoot && isNotIntrospection && keyDirectives.length > 0;
+
+  if (!check) {
+    return false;
+  }
+
+  const resolvableKeyDirectives = keyDirectives.filter(d => {
+    for (const arg of d.arguments) {
+      if (arg.name.value === 'resolvable' && arg.value.kind === 'BooleanValue' && arg.value.value === false) {
+        return false;
+      }
+    }
+    return true;
+  });
+
+  return { resolvableKeyDirectives };
 }
 
 /**